Moved Fragment classes to fragments package
